requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
sentence-transformers==3.1.1
langchain==0.3.1
langchain-openai==0.2.1
langchainhub==0.1.21
transformers==4.44.2
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.3
pyyaml==6.0.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9